// src/interface/user.d.ts
// export interface User {
//   uid: string;
//   displayName?: string | null;
//   email?: string | null;
//   photoURL?: string | null;
//   emailVerified: boolean;
//   phoneNumber?: string | null;
//   isAnonymous: boolean;
//   tenantId?: string | null;
// }

export interface IUser {
  id: string;
  username: string;
  bio: string;
  profile: string;
}